Update DTOs and entities except application.properties

+++ spring-project/src/main/java/ktc/spring_project/entities/Status.java
@@ -12,7 +12,7 @@ public class Status {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Short id;
 
     @Column(length = 50)
     private String type;
@@ -35,11 +35,11 @@ public class Status {
 
     // Getters and setters
 
-    public Long getId() {
+    public Short getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Short id) {
         this.id = id;
     }
